Correct Hmac key container implementation (#21)
As specified in RFC 2104, if key is shorter than the digest algorithm’s block length, then it will be padded with zeros. Similarly, if it
is longer than the block length then it will be compressed using the digest algorithm.

So this commit change to not use Tag as container for Hmac key and instead just use Vec<u8>