add support of custom RNG
Current two RNG we provided in this crate does not
meet FIPS requirements. So, this commit updates
all crypto algorithms that need to use RNG to have
a function pointer for creating RNG. In this way,
user could customize the RNG they want to use for
each algorithm.

Other Changes:

- Add unit tests.
- Fix clippy warnings.